+++ core/src/board/boardType.ts
@@ -1,0 +1,3 @@
+import { NullableTile } from "../tile";
+
+export type Board = NullableTile[][];
