+++ core/src/board/index.ts
@@ -1,1 +1,3 @@
 export * from "./board";
+export * from "./boardType";
+export * from "./boardGetters";
