+++ core/src/game/state.ts
@@ -1,4 +1,5 @@
-import { Board, getLastPlayedTile } from "../board/board";
+import { getLastPlayedTile } from "../board/boardGetters";
+import { Board } from "../board/boardType";
 import { getPlayableTilesForNextMove } from "../move/move";
 import { BLACK_PLAYER, Player } from "../player/player";
